Refactored config handling + submodules update (fmt, json, fb2k_utils)

+++ foo_spider_monkey_panel/utils/gdi_error_helpers.h
@@ -38,7 +38,7 @@ void CheckGdiPlusObject( const std::unique_ptr<T>& obj, const T_Parent* pParentO
 
     if ( status )
     {
-        throw qwr::QwrException( "Failed to create GdiPlus object ({:#x}): {}", *status, GdiErrorCodeToText( *status ) );
+        throw qwr::QwrException( "Failed to create GdiPlus object ({:#x}): {}", static_cast<int>( *status ), GdiErrorCodeToText( *status ) );
     }
     else
     {
